+++ kong/plugins/ip-restriction/schema.lua
@@ -10,6 +10,8 @@ return {
         fields = {
           { allow = { type = "array", elements = typedefs.ip_or_cidr, }, },
           { deny = { type = "array", elements = typedefs.ip_or_cidr, }, },
+          { status = { type = "number", required = false } },
+          { message = { type = "string", required = false } },
         },
         shorthand_fields = {
           -- deprecated forms, to be removed in Kong 3.0